refactor(workflow): move models to root directory for clarity
This commit refactors the project structure by moving several
model files from the `apps.workflow.models` subdirectories to
the root `apps.workflow.models` directory. This change aims to
improve code discoverability and maintainability by centralizing
model definitions in a single, well-defined location.

The following files were moved:
- `apps/job/helpers.py`
- `apps/purchasing/services/quote_to_po_service.py`
- `apps/workflow/admin.py`
- `apps/workflow/api/reports/pnl.py`
- `apps/workflow/api/xero/reprocess_xero.py`
- `apps/workflow/api/xero/sync.py`
- `apps/workflow/api/xero/xero.py`
- `apps/workflow/helpers.py`
- `apps/workflow/management/commands/generate_kpi_test_data.py`
- `apps/workflow/management/commands/start_xero_sync.py`
- `apps/workflow/views/xero/xero_invoice_manager.py`
- `apps/workflow/views/xero/xero_quote_manager.py`
- `apps/workflow/views/xero/xero_view.py`
