+++ apps/workflow/views/xero/xero_quote_manager.py
@@ -12,8 +12,7 @@ from .xero_base_manager import XeroDocumentManager
 from .xero_helpers import format_date, parse_xero_api_error_message # Assuming format_date is needed
 
 # Import models
-from apps.workflow.models import Quote, Client
-from apps.job.models import Job
+from apps.workflow.models import Quote
 from apps.job.enums import QuoteStatus
 from xero_python.accounting.models import LineItem, Quote as XeroQuote
 from xero_python.exceptions import AccountingBadRequestException, ApiException # If specific exceptions handled
